feat: 통합테스트 대성공

+++ src/main/java/com/project/baedalsodae/payment/pg/dto/PGPaymentRequest.java
@@ -5,4 +5,4 @@ import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PGPaymentRequest(
-        UUID orderId, UUID userId, PaymentMethod paymentMethod, BigDecimal amount) {}
+        UUID orderId, UUID userId, PaymentMethod paymentMethod, long amount) {}
